+++ src/main/java/in/oxidane/work/done/order/repository/LineItemRepository.java
@@ -8,13 +8,13 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface LineItemRepository extends JpaRepository<LineItem, Integer> {
-    
+public interface LineItemRepository extends JpaRepository<LineItem, Long> {
+
     Optional<LineItem> findByLineItemHandle(String handle);
-    
+
     boolean existsByLineItemHandle(String handle);
-    
-    boolean existsByLineItemHandleAndLineItemIdNot(String handle, Integer id);
-    
-    List<LineItem> findByUnitOfMeasurement_UomId(Integer uomId);
-} 
+
+    boolean existsByLineItemHandleAndLineItemIdNot(String handle, Long id);
+
+    List<LineItem> findByUnitOfMeasurement_UomId(Long uomId);
+}
